@@ -4,32 +4,32 @@
   "private": false,
   "license": "MIT",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.4.0",
     "history": "4.10.1",
     "normalize.css": "^8.0.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-collapsible": "^2.8.3",
-    "react-dom": "^17.0.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-collapsible": "^2.10.0",
+    "react-dom": "^18.2.0",
     "react-js-pagination": "^3.0.3",
-    "react-lazy-load": "^3.1.13",
-    "react-loading-skeleton": "^2.1.1",
-    "react-modal-video": "^1.2.6",
-    "react-redux": "^7.2.2",
-    "react-responsive-modal": "^4.0.1",
+    "react-lazy-load": "^4.0.1",
+    "react-loading-skeleton": "^3.3.1",
+    "react-modal-video": "^2.0.0",
+    "react-redux": "^8.0.5",
+    "react-responsive-modal": "^6.4.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "react-slick": "^0.25.2",
+    "react-scripts": "5.0.1",
+    "react-slick": "^0.29.0",
     "react-star-ratings": "^2.3.0",
-    "react-toastify": "^7.0.3",
-    "redux": "^4.0.5",
+    "react-toastify": "^9.1.3",
+    "redux": "^4.2.1",
     "redux-persist": "^6.0.0",
-    "redux-saga": "^1.1.3",
+    "redux-saga": "^1.2.3",
     "slick-carousel": "^1.8.1",
-    "web-vitals": "^1.0.1",
+    "web-vitals": "^3.3.1",
     "webfontloader": "^1.6.28"
   },
   "scripts": {
@@ -58,8 +58,8 @@
   },
   "devDependencies": {
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
-    "enzyme-to-json": "^3.6.1",
-    "node-sass": "4.14.1"
+    "enzyme-adapter-react-16": "^1.15.7",
+    "enzyme-to-json": "^3.6.2",
+    "sass": "^1.62.1"
   }
 }
