@@ -1,11 +1,10 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import { Provider } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
 import WebFont from "webfontloader";
 
 import { store, persistor } from "./redux/store";
-import reportWebVitals from "./reportWebVitals";
 import App from "./App";
 import LoadingScreen from "./components/shared/LoadingScreen";
 import "normalize.css/normalize.css";
@@ -18,19 +17,16 @@ WebFont.load({
   },
 });
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <PersistGate loading={<LoadingScreen />} persistor={persistor}>
         <App />
       </PersistGate>
     </Provider>
-  </React.StrictMode>,
-
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
